proposal for a very simple connection pool
Client will now set up a pool of connections when created.

Each connection checks its ES server periodically. If the ES server
does not return HTTP status 200 when checked, the connection
is marked as broken. Broken connections are skipped by the pool.

The connection pool has a `GetNextRequestURL` method. It returns
the (base) URL of the Elasticsearch server to use for the next request.
Broken connections are skipped. If `GetNextRequestURL` cannot return
any connection, it will return `ErrNoClient`.

The connection pool currently uses a very simple round-robin to
select the next connection.