SearchService accepts SearchSource and raw interface{}
SearchService now has two methods which allows users to create their own
requests manually. One method of doing so is to set up the JSON to pass
as a search request as an interface{} and use SearchService.Source().
Another method is to create a new instance of SearchSource,
set up the request via its setters, then pass it to
SearchService.SearchSource().

Example usage can be found in the tests TestSearchSource and
TestSearchSearchSource.

Closes #20 and #25.