Add cluster sniffing, round-robin connection handling etc.
Up until now, you passed a list of URLs to `NewClient`. The client would
then keep on this list of URLs and hang on the one that works. When the
active node stops working, it will select a new one from the list passed
into `NewClient` and hang on that one again.

Although quite simple, it has a few trade-offs. First, a client always
uses the same node in the cluster. Second, a client does not
automatically find other available nodes in the cluster. Third, the
caller needs to know all nodes in the cluster in advance.

This commit brings several changes.

First, a client now has two background procedures that run
periodically: The sniffer and the health check.

The sniffing process detects all nodes in a cluster with the Nodes Info
API. The sniffing process is run periodically to find new nodes in a
cluster and remove those that have e.g. been shut down. This makes the
client aware of changes in your cluster. The default schedule for
sniffing is 15 minutes, but it can be changed with `SetSnifferSchedule`.

The health check process on the other hand pings all nodes detected in
the sniffing process, periodically. It marks connections to those nodes
as either dead or alive. Health checks are run in a shorter interval
than the sniffer. The default is 60 seconds, but it can be changed with
`SetHealthcheckSchedule`.

On top of that, you can now automatically retry failed requests
(disabled by default). Use `SetMaxRetries` to set the maximum number of
retries.

When requests fail, the connection to that node is marked as dead. Dead
connections are not used for requests. They might be set alive by a
health check later on. When a request succeeds, the connection to that
node is marked as healthy.

Requests will now be routed to all nodes in the cluster in a round-robin
fashion.

All in all, this should make Elastic aware of changes in the cluster,
use all nodes of the cluster in a fair fashion, and--with retries
enabled--should keep a client more robust to node failures.