Add Reindexer service
The Reindexer simplifies reindexing data in Elasticsearch. One can copy
the contents of one index to another, even in another cluster.

Indexing is basically a scan-and-scroll operation on the source side and
a bulk indexing operation on the target side.

As long as reindexing isn't an integral part of Elasticsearch, this
helper might be beneficial when you want to apply changes to mappings.