Configurable timeouts
Introduce configurable timeouts for health checks and sniffing. There
are two types of timeouts: those used on startup (used when creating a
new client) and those used for subsequent processes. The former help
with slow startup which seems likely on slow/lossy networks (issue #58
on GitHub has more details).

This commit also fixes a possible race condition at startup. Under
certain conditions we would have returned a new client without having
at least one active and healthy connection to Elasticsearch. An error
occurs when issuing the first request. At that time, the sniffer and
healthchecks are already started which might leave the client in a bad
state. So now, we check that we really have a good client in NewClient
and only return a client and no error when Elasticsearch is available.