Make reindexing work with child documents
Added a Fields method to ScanService that  allows the caller to speficy
the fields to load.

The Reindexer now requests the _parent and _router fields (on top of the default _source field).
When the reindexer finds those fields in the hit, it sets them on the IndexRequest.

In index_test:
Added a new 'comment' type that is a child of 'tweet' together with its mapping.
Added a comment instance to the test dataset.