Remove Reindexer.recreateTargetIndex
That method allowed the caller to specify the index settings for
number_of_shards and number_of_replicas, but otherwise prevented the
user from specifying a mapping.

Removed the method and the two settings.
Now the caller must setup the target index any way he likes before
invoking the reindexer (this is in conformity with the reindexer
package comment)