Add SearchSource and friends to Scan service
The Scan service provides a service-side cursor to the results of a
search request. However, it didn't have all the search features that
the Search service has. This commit adds them to Scan service,
especially SearchSource.

The Scan service is a bit off as it should definitely use
`search_type=scan` under its covers. However, that means that sorting
is out of context for Scan service (as `scan` instructs Elasticsearch
to disable sorting). However, the current behavior of Scan is to only
add `search_type=scan` when there is no sort order specified, and we
will keep this promise in v2. This should probably change in v3 though.