Add SearchSource and friends to Scan service
The Scan service provides a service-side cursor to the results of a
search request. However, it didn't have all the search features that
the Search service has. This commit adds them to Scan service,
especially `SearchSource`. It also adds some convenience functions to
directly access features of `SearchSource` in Scan and Search service,
e.g. `FetchSource` and `FetchSourceContext`.

The Scan and Scroll services should probably be merged into a single
iterator-like service for 3.0 final.