Retry only when HTTP request fails
`PerformRequest` retried even on valid responses, e.g. when ES
returned a 404. However, 404 can be a valid response from Elasticsearch
for certain services, e.g. the Exists service. Now we only
retry when `HttpRequest.Do` returns an error.

This commit also refactors error handling for services that return a
404 as normal behavior, e.g. `ExistsService`, `IndicesExistsService`
and `IndicesExistsTypeService`. The error code is now handled in the
service, not in the generic error handler (`errors.go`). Instead, one
can pass HTTP error codes to ignore now to the error handler and to
`PerformRequest`.