Prevent deadlock when all nodes are dead
If we don't rely on sniffing to find healthy nodes, a client might be
stuck with all nodes dead and never return.

To resolve this situation we mark all nodes as alive in that case.
The next call to `PerformRequest` will then find a node that is
marked as alive (not healthy) and will retry. If a node came back
in the meantime, it will be marked as healthy.

Such a deadlock situation will also get logged.

See discussion in #192.