Change Field Stats API
The Field Stats API is now in line with the code that the API code
generate produces. All fields specified in the REST API spec should be
there. Tests are fixed accordingly.

The `MinValue` field may not always be a string, but the
`MinValueAsString` value is. Same for `MaxValue` and
`MaxValueAsString`.

The comparison values are not always strings. They could be e.g.
float64, int64, dates etc.

See #292