Refactor ScrollService
`ScrollService` is now compatible to `ScanService`. `ScanService` will
be removed in Elastic 5.x (for Elasticsearch 5.x). It was a mistake to
begin with and has a data race if you use it with multiple goroutines.

This commit also specifies that `elastic.EOS` (end-of-stream) is the
same as `io.EOF`. You can use them interchangeably. The `elastic.EOS`
will also be removed in Elastic 5.x.