Fix and clarify scroll issues
This commit fixes the `ScrollService` so that it uses the complete
output of its configured `SearchSource`, not only its query.

We also clarify the example code for scrolling in that we pass
`json.RawMessage` now instead of a pointer to a `SearchHit`. The latter
might cause problems when being used with multiple goroutines.