Backport Retrier
This is a backport of the Retrier from v5, minus the `Context`.

The Retrier allows users to intercept failed requests. The Retrier
interface specifies a single Retry func with the following
signature:

    Retry(
      // ctx context.Context, // only in v3 and v5
      retry int,
      request *http.Request,
      response *http.Response,
      err error,
    ) (time.Duration, bool, error)

Retry can inspect the parameters to decide whether to retry or not.
E.g. the retry parameter is the number of retries already processed.
Notice that request and response may be nil. The err parameter is the
error returned from Elasticsearch or the HTTP client.

If elastic should retry calling Elasticsearch, Retry must return a
duration to wait for the next request and true. It can use a
backoff strategy to do so (see below).

If elastic should stop calling Elasticsearch, Retry must return
false as the 2nd return value.

If Retry wants to stop and return a custom error, it can do so
by returning a non-nil error as the 3rd return value.

To simplify the implementation of custom Retriers, elastic comes with
a number of backoff implementations (see backoff.go). There is e.g.
an ExponentialBackoff or a ConstantBackoff implementation.

See the project wiki for further information about Retriers.

See #495