Return response on Delete API with 404
When ES is asked to delete a document that does not exist, it returns
HTTP status code 404, but also returns a response. This response
contains additional information about the request and the failure.

Up until now, this response has been swallowed. This commit changes the
behavior in that it returns both an error AND a response in case of a
404, just like ES does.

This should not break existing clients as the error is still there. The
only change is that Delete now also returns the serialized response for
that edge case.

See #555