Add Tasks Get API; add async reindexing
This commit adds to PR #550 and changes a few things.

First, the Reindex.Start method is renamed to DoAsync to reflect that
Do starts the request. Reindex.DoAsync starts reindexing in the
background and returns a task id, which can be watched via the Task Get
API, which was also added in #550.

Furthermore, it adds some tests and missing fields in the response
types, and fixes a typo or two in the JSON structs.

See #550