Catch request cancelation for an edge case
In certain cases, the returned error on a canceled or deadlined
request is not `context.Canceled` or `context.DeadlineExceeded`,
but a `*url.Error` whose `Error` field carries the above context errors.

In the standard library, there is a specific test for redirects that
checks this case (https://golang.org/src/net/http/client_test.go#L329),
so we fix this in the same way in `PerformRequest`.

See #484