Change scripting
Scripting changed quite a bit in 6.0. First, we now only have scripting
types "inline" and "stored". For inline scripting code, we need to use
the `"source"` property now instead of `"inline"`, and there is no
shortcut to pass the code. For stored scripts, e.g. those stored in the
configuration, we need to use `"id"`. The `"file"` property has been
removed in 6.0.

For details, see
https://www.elastic.co/guide/en/elasticsearch/reference/6.0/modules-scripting-using.html.