Fix Task List API to accept strings as TaskIds
In the Java source, task identification is represented by the `TaskId`
class. That class contains a `nodeId` (a `string`) and a numerical `id`
(a `long`). In the Tasks API, these two get mixed. Sometimes the
`TaskId` is used, which is a concatenation of `nodeId`, a colon, and
the numerical `id`, which yields a `string`. Sometimes, though,
especially in the response structures, the `nodeId` and `id` get
returned individually. So we must look at each of the response
structures in the Java source to find out the correct type used in the
response.

This commit now changes the Task List API to accept task ids as a
string, because that's what the REST API spec defines. We also add the
`GroupBy` property, and some missing fields in the response structures.

We also update Travis configuration and `run-es.sh` to Elasticsearch
5.6.6.