config: Do not fail with empty index name
When using `config.Parse(...)`, we no longer fail with an empty index
name. It is perfectly valid from an application perspective to set up
an Elasticsearch configuration, and make the index name an application
concern. This is especially true when 7.x comes out, where types are
removed and it is typicaly for Elastic to be used with multiple indices.