Add cat API endpoints
This commit adds some cat APIs. They are not strictly necessary as they
only represents a terse look into (mostly) the Cluster APIs. However,
one might use them for convenience.

This is a more complex implementation than #757. However, it returns
data as JSON and is therefore useable in a programming environment.

Close #757