Support X-Pack API
We now have experimental support for the X-Pack APIs, starting
with the Watcher API.

See:
* https://www.elastic.co/guide/en/elasticsearch/reference/6.4/xpack-api.html
* https://www.elastic.co/guide/en/elasticsearch/reference/6.4/watcher-api.html

This commit also switches to docker compose for testing, running two
containers: One is the OSS version of ES, available on
`http://localhost:9200`. The other is the ex-Platinum version of ES,
available on `http://localhost:9210`.

Close #476