Dial functions; import path changes for Go 1.11
This commit prepares the 6.2.0 release. It comes with the following
set of changes.

First, we fix the import path to work correctly with Go modules
introduced in Go 1.11. Notice that this change requires using
Go 1.10.4+ or later. If you're on an earlier version of Go, consider
updating or use `6.1.x` until you can.

Second, we add a set of `Dial` functions, i.e. `Dial`, `DialContext`,
and `DialWithConfig` to better align with the standard library in terms
of keeping a long-running TCP connection with a resource, in our case:
Elasticsearch.

In v6, the `Dial` functions will be kept as an alias to `NewClient`.
This might change in v7.

Also, the `DialContext` and `DialWithConfig` functions allow to pass a
context to be able to cancel connecting early. Notice that cancelation
can happen both due to the context being canceled, or because of a
configured timeout, e.g. via `SetHealthcheckTimeoutStartup`, whichever
timeout comes first (close #815).