Revert changes to import path
The change to use Go modules was too early for most users. So this
commit reverts those changes.

From now on, we will also test with `GO111MODULE=on` as well as
`GO111MODULE=auto` on TravisCI to make sure we test for early adopters
as well as users not being able or willing to use Go modules now.