Remove /v6 from import path
This commit removes the `/v6` from the import path and reverts the Go
module from `github.com/olivere/elastic/v6` to just
`github.com/olivere/elastic`.

The reason for that is that existing tools like `dep` and `glide`
currently don't support minimal module-awareness and are therefor
confused with the trailing `/v6`. Go 1.9.7, Go 1.10.4 and Go 1.11 have
this kind of module-awareness now, but tooling lacks a bit behind.

This makes it impossible for Elastic to be a good citizen for both
camps, and we need to delay 100% Go modules support to a later release.