Fix HTTP headers in the Tasks API
This commit fixes an issue with the Tasks API. When passing a HTTP
header in the request, it couldn't be deserialized in the HTTP response.

Furthermore, we now add a feature to retrieve the HTTP header from the
HTTP response header, not from its body as well. This is described
here:
https://www.elastic.co/guide/en/elasticsearch/reference/6.4/tasks.html#_identifying_running_tasks.

Close #906