Add ability to specify filter path to the Bulk service
We can already do it for the Search and Scroll services,
and as per https://www.elastic.co/guide/en/elasticsearch/reference/current/common-options.html
all REST APIs support specifying the filter path.

We would want to do this since if we are bulking a large amount of requests,
all of them needs to be serialized and passed over the network even if
we do not care about the result.