Use json.RawMessage instead of *json.RawMessage
This is a long-standing issue which we can remove now in v7.

It is not necessary to use `*json.RawMessage` to access to `Source` of a
search hit. We can simply use `json.RawMessage`, as it's a `[]byte`, so
the `*json.RawMessage` was a `*[]byte`.

This simplifies the code but is a breaking change, but that's okay for a
major version.

Close #687