Add rest_total_hits_as_int to Search and Msearch API
https://www.elastic.co/guide/en/elasticsearch/reference/7.0/breaking-changes-7.0.html#hits-total-now-object-search-response

To aid migration from 6.x to 7.x.

In 7.x hits.total in the search response is an object instead of an int.

rest_total_hits_as_int can be used to tell a 7.x server to return hits.total
as a number so that it is compatible with the 6.x search response format.