Add support for rest_total_hits_as_int
ES 7.x has added support for returning hits total as an integer rather
than a structure. This helps with updating from ES 6.x to ES 7.x, but
will be removed in ES 8.x.

See
https://www.elastic.co/guide/en/elasticsearch/reference/7.0/breaking-changes-7.0.html#hits-total-now-object-search-response.

Notice that enabling this flag indicates a usage scenario that is
unsupported: elastic.v6 has to be used with ES 6.x, and elastic.v7 has
to be used with ES 7.x. So use it at your own risk.

Close #1007