Find missing fields and deprecations
This commit adds two arguments that can be passed when running the test
suite.

First, `strict-decoder` is a boolean flag that enables a strict
JSON decoder that logs all fields that get serialized by Elasticsearch
but do not have a corresponding field in the library.

Second, `deprecations` can be set to `on` for logging or `error`/`fail`
to fail the given test when a deprecation warning is reported by
Elasticsearch in the HTTP response header.

Example:

  go test -race -deprecations -strict-decoder -v .

It also fixes the issues found with `strict-decoder` enabled.
Deprecations get logged in CI, but don't cause the build to fail.