Cleanup code
This commit adds three arguments that can be passed when running the test
suite.

First, `strict-decoder` is a boolean flag that enables a strict
JSON decoder that logs all fields that get serialized by Elasticsearch
but do not have a corresponding field in the library.

Second, `deprecations` can be set to `on` for logging or `error`/`fail`
to fail the given test when a deprecation warning is reported by
Elasticsearch in the HTTP response header.

Third, the `types-removal=true|false` (default is `false`) ensure that
deprecation warnings due to types removal in 7.x should be ignored.

Example:

   go test -race -deprecations -strict-decoder -types-removal=false -v .

This commit also cleans up a few warnings by staticcheck.

We also added a few missing fields in the structs found by the newly
added strict decoder.

Close #1183