Set XPackWatchActionStatus members to type XPackWatchActionExecutionState
XPackWatchActionStatus.LastExecution and LastSuccessfulExecution members
were of type `*time.Time`, but Elastic returns an object here. This was
the opposite case with XPackWatchStatus.ExecutionState which we changed
to a string because Elastic emits strings in #1211.

Without this change we get a parse error:
```
parsing time \"{\"timestamp\":\"2019-10-22T15:07:37.519Z\",\"successful\":true}\" as \"\"2006-01-02T15:04:05Z07:00\"\": cannot parse \"{\"timestamp\":\"2019-10-22T15:07:37.519Z\",\"successful\":true}\" as \"\"\"
```