Clean up, GitHub actions, Go update, and more
With today's release of Go 1.14, we officially support Go 1.13 and 1.14
and changed the tests accordingly.

We also updated the `go.mod` file to be based on Go 1.14.

We removed license checks with Fossa because we actually don't require it.

We updated the dependencies and re-generated the easyjson files with the
new version.

We fixed a number of issues reported by staticcheck.

We fixed a test on which ES reported duplicate header fields.

We added missing header text to some files where it was missing.

We switched from Travis to GitHub actions for testing, and it seems to
be working okay. We kept the `.travis.yml` file for now so we can easily
switch back.

We added `goreleaser` so that it automatically writes a CHANGELOG on
releases.