Fix TotalHits to work with both total hits as integers and struct
ES 7.x changed the way to return total hits from a simple integer value
to a structure with value and relation fields. The `TotalHits()` method
on the search response didn't take this into account. Now it does, and
`TotalHits()` works with both `RestTotalHitsAsInt(true)` and
`RestTotalHitsAsInt(false)` (the default).

Close #1298