Handle HTTP status 401 Unauthorized
This commit adds a way to check if the client failed to setup
due to HTTP status 401 Unauthorized. This may happen if you protect your
cluster with X-Pack Security.

The canonical way to do so is by checking the returned error from
`elastic.NewClient` with `elastic.IsUnauthorized(err)`, which returns
`true` or `false`.

Here's an example:

```
package main

import (
	"fmt"
	"net/http"

	elastic "github.com/olivere/elastic/v7"
)

func main() {
	_, err := elastic.NewClient(
		elastic.SetURL("http://127.0.0.1:9200"),
		elastic.SetBasicAuth("invalid-user", "invalid-password"),
		elastic.SetSniff(false),
	)
	switch {
	case elastic.IsUnauthorized(err):
		fmt.Printf("unauthorized")
	case elastic.IsStatusCode(err, http.StatusUnauthorized): // works as well
		fmt.Printf("authentication failed")
	case err != nil:
		panic(fmt.Sprintf("client: %v", err))
	}
}

```

Close #1357