Add new index template APIs
This commit implements the new Index Template APIs as described in
https://www.elastic.co/guide/en/elasticsearch/reference/7.9/index-templates.html.

They replace the legacy/v1 APIs and were implemented in 7.8.

In general, use `client.IndexXXXIndexTemplate(...)` instead of
`client.IndexXXXTemplate(...)` (with `XXX` being `Get`, `Put`, or
`Delete`) to use the newer versions.

Notice that this commit only implements the new Index Template APIs, not
the Component Template APIs.

Close #1393