Add option to close idle connections for dead nodes
This commit adds a configuration option `SetCloseIdleConnections` to a
client. The effect of enabling it is that whenever the Client finds a
dead node, it will call `CloseIdleConnections` on the underlying HTTP
transport.

This is useful for.e.g. AWS Elasticsearch Service. When AWS ES
reconfigures the cluster, it may change the underlying IP addresses
while keeping the DNS entry stable. If the Client would _not_ close idle
connections, the underlying HTTP client would re-use existing HTTP
connections and use the old IP addresses. See #1091 for a discussion of
this problem.

The commit also illustrates how to connect to an AWS ES cluster in the
recipes in
[`recipes/aws-mapping-v4`](https://github.com/olivere/elastic/tree/release-branch.v7/recipes/aws-mapping-v4)
and
[`recipts/aws-es-client`](https://github.com/olivere/elastic/tree/release-branch.v7/recipes/aws-es-client).
See the `ConnectToAWS` method for a blueprint of how to connect to an
AWS ES cluster.

See #1091