Update Nodes Stats API to latest API
This commit brings the Nodes Stats API to the latest version.

Notice that 7.15.0 has a
https://github.com/elastic/elasticsearch/issues/78311 that is likely
being fixed by https://github.com/elastic/elasticsearch/pull/78531, so
we're skipping the given test for that specific release.

Once 7.15.1 is released, the bug might reappear unless the PR is being
merged into that version.

See #1535