Fix: Bulk processor retries indefinitely on failure
When all retries are exhausted the worker internal requests buffer needs to be
cleared in failure scenarios. That is required because the commitFunc (and
consequently the underlying BulkService.Do call) doesn't reset it when some error
happens. Without clearing the internal buffer the worker will continue sending
the same requests on the following rounds of execution.

Kudos for this solution goes to @rwynn and @raiRaiyan .

Resolves #1278