fix: attempt to retry and mark connection as dead
When nodes are being shutdown quickly we have a window when
a connection is alive in the pool, but erroring when a request is performed.

It does retry, however there is an edge case on the next pass a
connection could be chosen which should be dead.

This change alters the retry logic, when an error is returned from
the retrier and continue == true, it will mark the connection as dead
and retry.